Isolate a publically includable headers for RHI (names temporary for now)
